requirements.txt
Django==1.8.2
django-bootstrap3==5.4.0
Pillow==3.2.0
